Add intial index parameter to 'TitleView'
At the present, one can set an initial index with the use of a 'TableController'. Unfortunately, it is set privately in a 'DropdownController' and the 'dropdown' property set in the 'TitleView' cannot be changed because of it's reference in the 'action' block that would not be called properly if the 'dropdown' property is changed after initialization.

This was the easiest way to allow the 'initial index' functionality at the present.